[editor] Split the results grid and charts across two tabs
