+++ desktop/libs/notebook/src/notebook/static/notebook/less/editor2.less
@@ -115,34 +115,102 @@
 
               overflow: auto;
             }
+          }
 
+          .editor-chart-panel {
             .result-tab-body {
-              .flexRowLayoutRow(0 1 100%);
+              .flex(0 1 100%);
 
               overflow: hidden;
+            }
 
-              .table-results {
-                .fillAbsolute();
+            .table-results,
+            .chart-container {
+              height: 100%;
+              width: 100%;
+              position: relative;;
+            }
 
-                .table-results-inner,
-                .table-results-inner > .table-results-bridge {
-                  position: relative;
-                  height: 100%;
-                  width: 100%;
+            .result-settings {
+              .nav-list .nav-header {
+                margin-bottom: -3px;
+                padding-left: 3px;
+              }
+
+              .checkbox {
+                padding-top: 0;
+              }
+            }
+
+            .split-result-container {
+              .display-flex();
+
+              position: relative;
+              width: 100%;
+              height: 100%;
+
+              .result-settings-panel {
+                .flex(0 0 200px);
+
+                min-width: 200px;
+                overflow-y: auto;
+                padding-bottom: 15px;
+              }
+
+              .split-result-resizer {
+                .flex(0 0 3px);
+
+                cursor: ew-resize;
+              }
+
+              .split-result-content {
+                .flex(0 1 100%);
+
+                overflow-y: auto;
+
+                .resultTable th {
+                  color: @cui-white;
                 }
 
-                .table-results-inner .hue-table-container {
-                  height: 100%;
+                .dataTables_wrapper .resultTable th {
+                  color: @cui-gray-800;
+                }
+
+                .results-images {
+                  border-left: 5px solid @cui-default-border-color;
+                  background-color: @cui-gray-050;
+                  padding: 5px;
+                }
+
+                .table-results {
+                  position: relative;
                   width: 100%;
-                  overflow: auto;
+                  min-height: 250px;
                 }
               }
+            }
+          }
+          .execution-results-tab-panel {
+            .fillAbsolute();
+
+            .table-results-bridge,
+            .hue-table-container {
+              position: relative;
+              height: 100%;
+              width: 100%;
+            }
+
+            .hue-table-container {
+              overflow: auto;
             }
           }
         }
       }
     }
   }
+
+
+
 
 
   .snippet-move-placeholder {
@@ -411,75 +479,6 @@
     .btn-editor:last-of-type {
       border-top-right-radius: 2px;
       border-bottom-right-radius: 2px;
-    }
-  }
-
-  .snippet-tab-body {
-    padding: 0 5px;
-
-    .snippet-tab-actions-append {
-      position: absolute;
-      left: 150px;
-      top: -35px;
-    }
-
-    .result-settings {
-      .nav-list .nav-header {
-        margin-bottom: -3px;
-        padding-left: 3px;
-      }
-
-      .checkbox {
-        padding-top: 0;
-      }
-    }
-
-    .split-result-container {
-      .display-flex();
-
-      position: relative;
-      width: 100%;
-
-      .result-settings-panel {
-        .flex(0 0 200px);
-
-        min-width: 200px;
-      }
-
-      .split-result-resizer {
-        .flex(0 0 3px);
-
-        cursor: ew-resize;
-      }
-
-      .split-result-content {
-        .flex(1);
-
-        .resultTable th {
-          color: @cui-white;
-        }
-
-        .dataTables_wrapper .resultTable th {
-          color: @cui-gray-800;
-        }
-
-        .results-images {
-          border-left: 5px solid @cui-default-border-color;
-          background-color: @cui-gray-050;
-          padding: 5px;
-        }
-
-        .table-results {
-          position: relative;
-          width: 100%;
-          min-height: 290px;
-
-          table > thead > tr > td:first-child,
-          table > tbody > tr > td:first-child {
-            background-color: @cui-gray-050;
-          }
-        }
-      }
     }
   }
